requirements.txt
Django==1.7
djangorestframework==3.3.3
djangorestframework-bulk==0.2.1
tinyobj==0.1.0
